requirements.txt
Django==3.0.8
djangorestframework==3.11.0
Markdown==3.2.2
django-filter==2.3.0
powerfuldeveloper-psutils==0.0.4